moving ripgrep stuff to new location
