+++ lua/grug-far/actions/replace.lua
@@ -1,6 +1,6 @@
-local fetchFilesWithMatches = require('grug-far/rg/fetchFilesWithMatches')
-local fetchReplacedFileContent = require('grug-far/rg/fetchReplacedFileContent')
-local getArgs = require('grug-far/rg/getArgs')
+local fetchFilesWithMatches = require('grug-far/engine/ripgrep/fetchFilesWithMatches')
+local fetchReplacedFileContent = require('grug-far/engine/ripgrep/fetchReplacedFileContent')
+local getArgs = require('grug-far/engine/ripgrep/getArgs')
 local renderResultsHeader = require('grug-far/render/resultsHeader')
 local resultsList = require('grug-far/render/resultsList')
 local utils = require('grug-far/utils')
