+++ lua/grug-far/engine/ripgrep/colors.lua
