+++ lua/grug-far/engine/ripgrep/parseResults.lua
@@ -1,4 +1,4 @@
-local colors = require('grug-far/rg/colors')
+local colors = require('grug-far/engine/ripgrep/colors')
 local is_win = vim.api.nvim_call_function('has', { 'win32' }) == 1
 
 --- @enum ResultsTokenType
